+++ codeup-20200414/ChangeLog.md
@@ -1,3 +1,7 @@
+2021-09-03 Version: 0.1.6
+- Support query organization repository settings.
+- Support trigger mirror sync openapi.
+
 2021-08-25 Version: 0.1.5
 - Support delete repo member with yunxiao uid.